+++ lib/Google/Ads/GoogleAds/V19/Enums/LocalServicesLeadSurveyAnswerEnum.pm
@@ -12,25 +12,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-package Google::Ads::GoogleAds::V19::Resources::ImprovePerformanceMaxAdStrengthRecommendation;
+package Google::Ads::GoogleAds::V19::Enums::LocalServicesLeadSurveyAnswerEnum;
 
 use strict;
 use warnings;
-use base qw(Google::Ads::GoogleAds::BaseEntity);
 
-use Google::Ads::GoogleAds::Utils::GoogleAdsHelper;
-
-sub new {
-  my ($class, $args) = @_;
-  my $self = {
-    adStrength => $args->{adStrength},
-    assetGroup => $args->{assetGroup}};
-
-  # Delete the unassigned fields in this object for a more concise JSON payload
-  remove_unassigned_fields($self, $args);
-
-  bless $self, $class;
-  return $self;
-}
+use Const::Exporter enums => [
+  UNSPECIFIED       => "UNSPECIFIED",
+  UNKNOWN           => "UNKNOWN",
+  VERY_SATISFIED    => "VERY_SATISFIED",
+  SATISFIED         => "SATISFIED",
+  NEUTRAL           => "NEUTRAL",
+  DISSATISFIED      => "DISSATISFIED",
+  VERY_DISSATISFIED => "VERY_DISSATISFIED"
+];
 
 1;