+++ lib/Google/Ads/GoogleAds/V19/Enums/LocalServicesLeadSurveySatisfiedReasonEnum.pm
@@ -12,25 +12,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-package Google::Ads::GoogleAds::V19::Resources::ImprovePerformanceMaxAdStrengthRecommendation;
+package Google::Ads::GoogleAds::V19::Enums::LocalServicesLeadSurveySatisfiedReasonEnum;
 
 use strict;
 use warnings;
-use base qw(Google::Ads::GoogleAds::BaseEntity);
 
-use Google::Ads::GoogleAds::Utils::GoogleAdsHelper;
-
-sub new {
-  my ($class, $args) = @_;
-  my $self = {
-    adStrength => $args->{adStrength},
-    assetGroup => $args->{assetGroup}};
-
-  # Delete the unassigned fields in this object for a more concise JSON payload
-  remove_unassigned_fields($self, $args);
-
-  bless $self, $class;
-  return $self;
-}
+use Const::Exporter enums => [
+  UNSPECIFIED            => "UNSPECIFIED",
+  UNKNOWN                => "UNKNOWN",
+  OTHER_SATISFIED_REASON => "OTHER_SATISFIED_REASON",
+  BOOKED_CUSTOMER        => "BOOKED_CUSTOMER",
+  LIKELY_BOOKED_CUSTOMER => "LIKELY_BOOKED_CUSTOMER",
+  SERVICE_RELATED        => "SERVICE_RELATED",
+  HIGH_VALUE_SERVICE     => "HIGH_VALUE_SERVICE"
+];
 
 1;