+++ lib/Google/Ads/GoogleAds/V19/Enums/TargetFrequencyTimeUnitEnum.pm
@@ -20,7 +20,8 @@ use warnings;
 use Const::Exporter enums => [
   UNSPECIFIED => "UNSPECIFIED",
   UNKNOWN     => "UNKNOWN",
-  WEEKLY      => "WEEKLY"
+  WEEKLY      => "WEEKLY",
+  MONTHLY     => "MONTHLY"
 ];
 
 1;