+++ lib/Google/Ads/GoogleAds/V19/Resources/PmaxCampaignSettings.pm
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-package Google::Ads::GoogleAds::V19::Resources::ImprovePerformanceMaxAdStrengthRecommendation;
+package Google::Ads::GoogleAds::V19::Resources::PmaxCampaignSettings;
 
 use strict;
 use warnings;
@@ -22,9 +22,7 @@ use Google::Ads::GoogleAds::Utils::GoogleAdsHelper;
 
 sub new {
   my ($class, $args) = @_;
-  my $self = {
-    adStrength => $args->{adStrength},
-    assetGroup => $args->{assetGroup}};
+  my $self = {brandTargetingOverrides => $args->{brandTargetingOverrides}};
 
   # Delete the unassigned fields in this object for a more concise JSON payload
   remove_unassigned_fields($self, $args);