+++ lib/Google/Ads/GoogleAds/V19/Services/ConversionUploadService/SessionAttributeKeyValuePair.pm
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-package Google::Ads::GoogleAds::V19::Resources::ImprovePerformanceMaxAdStrengthRecommendation;
+package Google::Ads::GoogleAds::V19::Services::ConversionUploadService::SessionAttributeKeyValuePair;
 
 use strict;
 use warnings;
@@ -23,8 +23,8 @@ use Google::Ads::GoogleAds::Utils::GoogleAdsHelper;
 sub new {
   my ($class, $args) = @_;
   my $self = {
-    adStrength => $args->{adStrength},
-    assetGroup => $args->{assetGroup}};
+    sessionAttributeKey   => $args->{sessionAttributeKey},
+    sessionAttributeValue => $args->{sessionAttributeValue}};
 
   # Delete the unassigned fields in this object for a more concise JSON payload
   remove_unassigned_fields($self, $args);