+++ lib/Google/Ads/GoogleAds/V19/Services/ReachPlanService/PlannedProduct.pm
@@ -25,6 +25,7 @@ sub new {
   my $self = {
     advancedProductTargeting => $args->{advancedProductTargeting},
     budgetMicros             => $args->{budgetMicros},
+    conversionRate           => $args->{conversionRate},
     plannableProductCode     => $args->{plannableProductCode}};
 
   # Delete the unassigned fields in this object for a more concise JSON payload